+++ scripts/faces/blocks_set_operations.py
@@ -18,7 +18,7 @@ box_red.color = (0.4, 0.1, 0.1)
 box_red.name = 'box_red'
 
 box_green = box_red.frame_mapping(vm.Frame3D(vm.Point3D(-0.4, 0, -0.1), vm.Vector3D(1, 0, 0),
-                          vm.Vector3D(0, 1, 0), vm.Vector3D(0, 0, 1)), 'new', copy=True)
+                          vm.Vector3D(0, 1, 0), vm.Vector3D(0, 0, 1)), 'new')
 
 box_green.color = (0.1, 1, 0.1)
 box_green.name = 'box_green'
@@ -29,12 +29,12 @@ shell1.alpha = 0.6
 vm.core.VolumeModel([shell1]).babylonjs()
 #
 box_blue = box.frame_mapping(vm.Frame3D(vm.Point3D(0.1, 0, 0), vm.Vector3D(1, 0, 0),
-                          vm.Vector3D(0, 1, 0), vm.Vector3D(0, 0, 1)), 'old', copy=True)
+                          vm.Vector3D(0, 1, 0), vm.Vector3D(0, 0, 1)), 'old')
 box_blue.color = (0.1, 0.1, 1)
 box_blue.name = 'box_blue'
 
 box_blue2 = box.frame_mapping(vm.Frame3D(vm.Point3D(0.3, 0, 0), vm.Vector3D(1, 0, 0),
-                          vm.Vector3D(0, 1.8, 0), vm.Vector3D(0, 0, 1)), 'old', copy=True)
+                          vm.Vector3D(0, 1.8, 0), vm.Vector3D(0, 0, 1)), 'old')
 box_blue2.color = (0.1, 0.1, 1)
 box_blue2.name = 'box_blue2'
 shell2 = box_blue.union(box_blue2)[0]
@@ -56,3 +56,27 @@ for new_box in [union_box, subtraction_box, subtraction_closedbox, intersection_
         shell.color = (1, 0.1, 0.1)
         shell.alpha = 0.6
     vm.core.VolumeModel(new_box + [shell1, shell2]).babylonjs()
+
+
+block1 = primitives3D.Block(frame=vm.Frame3D(origin=vm.Point3D(0.2, 0.6, 0.2),
+                                                  u=vm.Point3D(0.4, 0, 0),
+                                                  v=vm.Point3D(0, 1.2, 0),
+                                                  w=vm.Point3D(0, 0, 0.4)),
+                            color=(1, 0.2, 0.1), alpha=0.6)
+block2 = primitives3D.Block(frame=vm.Frame3D(origin=vm.Point3D(0.8+0.4, 0.45 + 0.15 , 0.1),
+                                                  u=vm.Point3D(1.6, 0, 0),
+                                                  v=vm.Point3D(0, 0.9, 0),
+                                                  w=vm.Point3D(0, 0, 0.2)),
+                            color=(0.2, 0.2, 0.2), alpha=0.6)
+
+block3 = primitives3D.Block(frame=vm.Frame3D(origin=vm.Point3D(0.8+0.4, 0.45 + 0.15 , 0.05),
+                                                  u=vm.Point3D(1.6, 0, 0),
+                                                  v=vm.Point3D(0, 0.9, 0),
+                                                  w=vm.Point3D(0, 0, 0.2)),
+                            color=(0.2, 0.2, 0.2), alpha=0.6)
+
+for battery_layout_max_enclosing_box in [block1.union(block2)[0], block1.union(block3)[0]]:
+    battery_layout_max_enclosing_box.merge_faces()
+    battery_layout_max_enclosing_box.color = (1, 0.2, 0.2)
+    battery_layout_max_enclosing_box.alpha = 0.6
+    battery_layout_max_enclosing_box.babylonjs()